requirements.txt
# write only package name which should be installed before testing (to build project successfully)
pytest
numpy
Pillow